+++ Tests/Functional/Service/QueueServiceTest.php
@@ -45,8 +45,7 @@ class QueueServiceTest extends FunctionalTestCase
         $crawlerController = $this->getAccessibleMock(CrawlerController::class, ['getCurrentTime'], [], '');
         $crawlerController->method('getCurrentTime')->willReturn(100000);
 
-        $this->subject = $this->createPartialMock(QueueService::class, []);
-        $this->subject->injectCrawlerController($crawlerController);
+        $this->subject = GeneralUtility::makeInstance(QueueService::class, $crawlerController);
         $this->queueRepository = GeneralUtility::makeInstance(QueueRepository::class);
     }